Renamed Config to GlobalConfig.
The trigger for this renaming was a quite questionable change in
clang-format, see https://reviews.llvm.org/D33932.  With that change, header
matching is always case-insensitive now, and there is no flag or option to
get the old behavior back.  Due to this, our "config.h" and "Config.h"
headers were considered the same during formatting, which is wrong.  :-/

Instead of fighting against an extremely useful tool, we simply rename the
class and files in question to the arguably better name "GlobalConfig".

Change-Id: I737a67c560a30a06c7d0bcea1cf82c439fcadc3a