5562 Flap settings can now be configured individually for hosts/services
It is now possible to configure the flaping settings individuall for each
host and service. This was previously only configurable globally via
the global setting "Tuning of flap detection".

No you can use the rulesets "Flap detection settings for hosts" and
"Flap detection settings for services" to configure this using rules.

Change-Id: I2137d9f4d9f7c7bb11a8407aea55f89ffb13ce45