5563 FIX Renamed "Site Management" configuration options introduced with #5173
The configuration options introduced with #5173 (Set OMD config via WATO) had
to be renamed. This is mostly an internal change, but you may have set one
of these options in a site using a daily master snapshot or the 1.5.0i1
version. Renaming the settings will reset your already configured settings
to default.

Change-Id: Id8c88d514d063f5a82488ae979fcc758233b3125