Windows agent: use RAII for event handles
More precisely, use the Resource Release is Finalization principle as
the handles cannot be always acquired in a uniform way in a constructor
due to the various ways handles can be acquired/opened through WinAPI.

Wrap handles related to eventlog handling in instances of classes
responsible for releasing the resources (alias closing the handles in
this case). Specialize based wrapper class through handle-specific
traits classes for handle type, invalid value as well as closing the
handle.

Do not wrap wrapper classes further to unique_ptr's etc when. there is no
need for heap allocation of the wrapper object.

Also refactor the eventlog handling in order to get rid of unnecessary
resetting.

Add unit tests for handle wrappers.

Change-Id: Iacd6e245d5e55b6276f3ce3509ea6679103e7e8b