Ignore the dreaded leaking file descriptor problem for now.
The problem of leaking file descriptors across fork()/exec() is a dark and
heavily disputed corner of many *nices, including Linux.  In a nutshell:
It's a horrible design flaw, and there are basically two approaches of
working around it:

   * First workaround: You directly open *all* your files with O_CLOEXEC,
     but to do that atomically, you need a bunch of new system calls.
     Naively using "fcntl(fd, F_SETFD, FD_CLOEXEC)" introduces a race
     condition during the time between open() and fcntl() where a
     fork()/exec() may happen.  In addition you'll have various kinds of fun
     trying to dup()/dup2()/dup3() your file descriptors correctly.

   * Second workaround: Explicitly close all file descriptors you don't need
     anymore before a fork()/exec().  This has various downsides, too: You
     don't know exactly which file descriptors are open (think: libraries!),
     and which of them should survive.  Furhtermore, you can only pray that
     your libraries either don't fork()/exec() or that they use the same
     workaround as you.

Note that both approaches only really work if all SW components use the same
workaround, which is *far* from reality.  Just a few entertaining references
to see that this is not a problem which is easy to solve and where everybody
agrees:

   * glibc bug 10353 "Methods for deleting all file descriptors greater than
     given integer" (https://sourceware.org/bugzilla/show_bug.cgi?id=10353):
     Several people request glibc support for the 2nd workaround, but Ulrich
     Drepper is as "friendly" and "open-minded" as usual and insists on the
     1st workaround, ignoring its inherent problems...  :-P

   * Python issue 21207 "urandom persistent fd - not re-openned after fd
     close" (https://bugs.python.org/issue21207): Some library uses the 2nd
     workaround, but Python is not prepared for that.

   * "Excuse me son, but your code is leaking !!!"
     (http://danwalsh.livejournal.com/53603.html): One of many blog entries
     advertising the first workaround.

   * "I'm done with O_CLOEXEC"
     (https://mail.gnome.org/archives/gtk-devel-list/2015-March/msg00038.html):
     A rant from a core GNOME developer, saying goodbye to the 1st
     workaround.

As a fun fact, ancient CentOS 5 ships with Linux 2.6.18 from 2006, but
O_CLOEXEC was introduced only in 2.6.23 from 2007.  Furthermore, the
corresponding "e" modifier for fopen() was introduced in glibc 2.7 (2007),
but CentOS 5 ships with the brand-new, bleeding-edge glibc version 2.5 from
2006.  :-/ So we can't use the 1st workaround yet, only when we drop support
for CentOS 5.

What the NEB/CMC currently does is a mixture of both workarounds, including
some hopefully harmless race conditions.  This should be good enough for
now, but we should definitely come back to that dark corner at some point in
time.

Change-Id: Ic7f78e13acc866cea81ec87bb92fc58cd812d3a1