Made counter for active Livestatus connections atomic.
We have seen Livestatus usage >100% out in the wild, which is total
nonsense.  This commit is trying to fix this, but there could be other
problems: We really, really have to introduce memory barriers and locks in
Livestatus, the current code is simply wrong in this respect.

This is a port of 863bb624d6667004a70878734a272cb5c70c9816.

Change-Id: I153389f4a12ceaea596d8bb889da57207803db8d