5482 Add option to specify a timeperiod for notification bulking
The option "Bulking period" can be used in notification rules to
specify a time period for which notifications should always be
bulked. After the timeperiod has ended the notifications of the
bulk will be sent. This option is useful if e.g. after the
working hours no notifications should be sent.

Change-Id: Ibe4dec64649d2a08f62fa8d7234c02cdd842d8db