5415 FIX Make Windows section openhardwaremonitor backwards compatible
Version 1.4.0 'corrected' the naming of section openhardwaremonitor.
Unfortunately, this broke the backwards compatibility of old ini files
that still call this 'ohm'. The fix maps the old name to the new one
so that both old and new ini files work. Moreover, the agent bakery
was still writing ini files with the old name. Now the agent bakery
is also fixed.

Change-Id: I5adf3680c46c39c5833b5e39073064b1e2c74f5d