5416 FIX Prevent MS Exchange checks from crashing upon WMI timeout
Werk #5411 introduced the identification and handling of WMI timeouts.
Some of nested subsections, at least those of MS Exchange, may under
certain circumstances produce an unprecedented order of output lines
suddenly ending in a WMI timeout. This scenario was not expected when
preparing #5411, causing the parsing of some MS Exchange subsections
to fail and the corresponding check to crash.

This fix ensures that a WMI timeout is properly parsed regardless of
where it appears in a subsection output. The corresponding checks no
longer crash but issue a PENDING/UNKNOWN result.

Change-Id: Ic43050a153a1dcb05286564d4a5478934b74fdd4