5476 FIX Fixed notification numbering of periodic notifications
Previously, non-OK state changes during periodic notifications resetted the
notification number back to 1, in effect restarting escalations.  This was
inconsistent with Nagios and is not what one wants: If e.g. a service is in
WARN state for quite some time, and then goes to CRIT, one surely wants to
continue escalation, not reset it.

This problem has been fixed, so problems correctly escalate now.

Change-Id: Ie4a6430619d1cb3548fa63349b4dc27e39d3e303