5486 Add permission for admins to see custom views and dashboards
Previously, admins could only see custom views, dashboards and reports of a
user if the owner set the option "Make this view/dashboard/report available for
other users". Now admins can see the views, dashboards and reports of all
users. This also implies that admins are allowed to edit and clone these views.

Change-Id: I4079adc1f960c5355ed4ca8e7d679049bbe7ad59