Windows agent: reduce logfile hint fields
Logfile hints are used to store offset information read from
logstate.txt. The hints were using the struct logwatch_textfile that
contains extra fields for this purpose and forced the use of a bogus
vector for patterns that never exist, among others. Reduce the number
of fields by introducing a reduced version of the struct
(logwatch_hint) that only contains the necessary fields.

Change-Id: I135d97616e90d80ea9685eb5ae2b6b406fb5ddcb