5631 New host/service painter to display the notification postponement reason of the Microcore
The Check_MK Microcore has a mechanic that delays notifications for different reasons. For example
service notifications are delayed when the service is in downtime or when the host of the service
is down.

With the new painters "Notification postponement reason" Check_MK is now showing the reason why
a pending notification for a host/service is not yet sent to the user. This painter has been
added to the host and service detail pages by default.

The users could already check the notification logic and see the current postponement reasons
using the cmc.log, but this needs command line access and is not very comfortable when having a
larger installation which produces a lot of log entries.

Change-Id: I8ab97dabe5d432932a3168baec6470ceabcade59