5632 SEC Fixed XSS when rendering values of dropdown choices
When using the WATO configuration it was possible to create e.g.
a service level definition with javascript code in it's alias. When
this definition was configured in a rule of the ruleset
"Service Level of Hosts", the javascript code could be executed in the
browsers context of the user viewing the rule.

The insertion of the javascript code is only possible for authenticated
users with the permission to configure Check_MK.

Change-Id: I968949787a22c30b63bcf3f4cf18a9a921d40770