5649 FIX Fixed performance issue caused by too many parse function calls
Fixed a regression that was introduced with 1.5.0i2. The parse functions
were called not only once per Check_MK check, as it is intended. Each
service of a Check_MK check lead to a new call of the parse function logic.

For example in case you have 100 services of a single check type on one host,
the parse logic was called a hundred times instead of a single time.

Change-Id: I66cb6c08393bab567f017c6e20f0cbf134b37173