5495 fileinfo: Make out of relevant time of day more present and remove state markers
Previously, the information that a service was out of the relevant time of day
was displayed at the end of the plugin output where it can be missed easily.
Now the information is displayed at the beginning of the plugin output.

Furthermore, the subresults of a service do not show {WARN} and {CRIT} markers
anymore since this yields the impression that the check should not be {OK}.
Additional information about violations are still shown, e.g. if the age
of a file is older than a configured level.

Change-Id: I5b6d068943b1acbe12b6c5915ed4b864aa80bf2b