5669 FIX Always fetch host/authorization-related info from EC when needed
To correctly show host-related information and to handle EC-related
authorization, Livestatus needs to retrieve enough information from the EC.
This was broken by an optimization in the master branch quite some time ago
(Oct 2017), so v1.5.0i1 and v1.5.0i2 showed some strange behaviour in this
area.  This is fixed again now.

Change-Id: I7473bc783c583d190a8413a031bf5d520c793986