5271 FIX Better handling of erroneous deployment information in Agent Bakery
Depending of the state of the agent deployment information file for a host, the pages "Automatic agent updates" and "Agent update status" would crash and show only partial information.

This is now mitigated by handling corrupt agent deployment files per-host and displaying an error message in a convenient format and location on the affected pages.

Change-Id: Id6ad62d2eafad6f17a1123c1871e75aa353effc5