5273 FIX Fix display of CustomGraphs for metrics without current performance data
The graphing system was unable to display metrics within a custom graph if the corresponding performance-data is currently not available. This would for example be the case if a check isin a temporary "Unknown" state.

The rendering now correctly relies on the corresponding historical metric data instead of discarding the metric completely.

Change-Id: I245a435350d87ced7adaea6f3e9c558258817f3d