5850 Management board protocol can now be configured on folders
To make the configuration more flexible it is now possible to
configure the management board protocol on folders and inherit
this option to the hosts in a folder.

It is also possible to override the value inherited from a folder
to e.g. set it to "No management board" for specific hosts again.

Change-Id: If9a163ecdeb331768ede45c484af76cdadb8b7b8