5742 windows_tasks: show the hexadecimal representation of the status code
schtasks.exe used by the check plugin windows_tasks.ps1 returns a signed
integer value as the last result of a task. However, the status codes
(e.g. on MSDN) are usually represented as a hexadecimal value. To make
it easier to lookup a status code the hexadecimal value is now shown by
the check plugin.

Change-Id: I591eb397c172446e7312cf43ed077227ae47cd47