5865 FIX omd rm --kill now kills all processes of the site user
Instead of just killing the processes that are currently accessing
the tmpfs of the site, the "omd rm" command together with the --kill
option now also kills all processes of the site user. This includes
open shells or similar that would otherwise make the site removal
fail in case such processes are open.

Change-Id: I88ec7cfdbf00f5f4950405d58138638c20480e64