Unconditionally import SNMP modules.
This saves only <200ms on startup when --snmptrap is not used, which is
basically nothing compared to loading the rest of the modules, the
configuration, and the status file.  Furthermore, it complicates two things:
Extracting global statements into a main() function and type checking via
mypy.  Both things could be worked around somehow, but this wouldn't exactly
result in the kind of code to be proud of...  :-}

Change-Id: I5743ab999f02505c80f01e802dbcbf72bd0f2ede