5872 FIX Do not bake VANILLA/GENERIC implicit when bake for list of hosts
When the agent bakery is invoked with "cmk -A HOST1" to bake the agents
for the host with the name HOST1, previous Check_Mk versions were baking
the VANILLA and GENERIC agents in addition to the given host. This has
been changed. Now the bakery only bakes the agents for the hosts the
user requested with the call.

Change-Id: I69c888a06bbd65a2b2f9711e8c95dd7cafb79afd