5785 FIX User notification rules: Now also checks if the user is restricted to authorized sites
If a user gets a restriction to authorized sites, it will also restrict the notifications to these sites.
Previously, if no </i>Match site</i> filter was set as specific condition, the user rule matched on all sites.

So, with this change a user notification rule is only checked on the users authorized sites.

Change-Id: I52065e4119075a70a91a9d68ee78dfd5df3ddce4