5909 FIX fortigate_sessions, fortigate_cpu: Fixed duplicated discovery if needed information is at two places
Fortigate changed obviously the location of information for cpu utilization
and sessions in the past. We support these by using two different check
plugins. With this fix the service discovery does not create duplicated
services anymore if both paths are present on the device. Instead the newer
path will be preferred.

Change-Id: I2c4bd51e629c47e14988ba6d4e8d83870b90d64f