Win-agent: Replace MultiByteToWideChar
Use std::wstring_convert::from_bytes instead of WinAPI
MultiByteToWideChar.

This was a TODO leftover item from the commit
df323514929740597f21db7c03460cf3ff28515e. There is a known bug in
MinGW-w64 / libstdc++.dll that results in incorrect endianness in the
conversion. With latest MinGW 5.0.3-1 based on gcc 7.2.0, we can use
the workaround of passing the desired endianness (little-endian) as a
template parameter to std::codecvt_utf8.