5751 FIX Add contact group contacts of EC rule if host is not a core host
When an event was received that did not belong to a host in the monitoring the
contacts of the contact groups specified in a Event Console rule were not
added to the notification context if the option "Host's contact groups have
precedence" was set in the Event Console rule. Now the contacts of the
specified contact groups are added even when the option is set.

Change-Id: I1a98fb4ddf5f6c2bca3022f6ff7946206e51a938