make dist: Cleanup false positive error messages
The excludes for the agent build directories were always applied while
the directories did not exist all the time. This made the tar command
exit with an error exit code (which was printed to the log, but ignored
by the makefile).

For better error handling the script would now fail in such a case. The
tar command has been fixed by making the excludes be applied optionally
only in case the directories in question exist.

Future todo: Make the whole Makefile use stricter shell operation mode.

Change-Id: Ieb3015fa1116990117bb1cc73ceb58da9d916801