5792 FIX Agent Encryption: No longer trying to decrypt all incoming tcp data when a "Encryption" rule with an arbitrary value was set
The agent output decryption function was incorrectly called for hosts with a specifically "Encryption" rule set.
In most scenarios the decryption failed, because the plain text was obviously not decrypted correctly.
The fallback of the decryption error was to use the plain text instead for for further processing - this was conincidentally the right bevhaviour.

However, sometimes the decryption succeeded, which lead to garbled agent output and missing agent sections.

Change-Id: I54b302812df00e8eb4d021bacd74bfa7bf80ac6f