5796 FIX Fixed race condition in counter computation if a host uses real time checks
There was a risk that the counter data for the Check_MK checks could get lost or reset
to an earlier state, when the host received regular updates through the realtime check mechanism.
This has been fixed. As a side effect, the counter data is now only saved back to disk when there were
actual changes.

Change-Id: I4e6cac0527fbecbc35e3e1209bbc802bc7163b27