5907 FIX CentOS/RedHat: Add explicit dependency to php-cli to fix possible broken PNP graphs
The previous Check_MK packages were not explicitly declaring the dependency
to the php-cli package which is needed to make Dokuwiki, NagVis and PNP4Nagios
work.

The dependency was only declared for php which normally implies php-cli.
Unfortunately there are third party packages that don't have this implicit
dependency.

To make Check_MK work out of the box with those PHP installations, we now added
php-cli as explicit dependency.

Change-Id: I20ff20e466c4e0048a5f47737fb68b9d274bcbf2