Win-agent: extract SectionHeader
This is the first step of splitting the Section class (hierarchy) to
smaller logical entities. The section header forms its own independent
part of section output and has its own parameters that are independent
from the rest of the section.

Define SectionHeader as a separator-character-templated class that is
used as a member variable of Section. Subsection header is implemented
by subclassing the SectionHeader. In order to use the templated
SectionHeader with std::unique_ptr, define a abstract base class
ISectionHeader that all templated SectionHeaders inherit from.

The next steps in splitting Section will be to extract SectionBody to
its own class hierarchy and to provide a factory for instantiating the
sections.